Handle missed predictions when stylifying.
It seems to happen occasionally for either extremely high or extremely low notes, presumably because those are absent or virtually absent from the training data, that the model fails to predict a velocity, or that it only predicts a very low velocity.

Rectify this by using the last predicted velocity for those notes. (Although it'd of course be better to train the model with more comprehensive data.)